requirements.txt
pyupbit>=0.2.33
pandas>=2.0.0
numpy>=1.24.0
loguru>=0.7.0
pyyaml>=6.0
ta>=0.11.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
